fix: todo page fixes

+++ src/controllers/todos/clearCompleted.js
@@ -2,7 +2,8 @@ import { Todo } from '../../entities/todo.js'
 
 const clearCompleted = async (ctx) => {
   try {
-    const todos = await Todo.deleteMany({ completed: true })
+    const userId = ctx.state.user.id
+    const todos = await Todo.deleteMany({ user: userId, completed: true })
 
     ctx.body = todos
   } catch (error) {
